refactor(db): extract stream monitor to src/db/streamMonitor.ts
Move all stream group and streamer CRUD functions and their types into their
own module. No cross-module dependencies; all public exports are unchanged.
